Disciplinas/tads/coleta_armazenamento_ddd/requirements.txt
pandas>=2.0.3
mysql-connector==2.2.9